graph_rag_course/capstone_enterprise_assistant/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
neo4j>=5.14.0
langchain>=0.0.334
langgraph>=0.0.19
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
spacy>=3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
pinecone-client>=2.2.4
